Fix CSS specificity rules. Added control over the UISceneNode children and added setHighlightOverRecursive. This allows us to inspect nested UISceneNodes. Fix hit-box bug when changing display.

+++ include/eepp/ui/css/stylesheet.hpp
@@ -60,7 +60,7 @@ class EE_API StyleSheet {
 
 	void removeAllWithoutMarker( const Uint32& marker );
 
-	void setSelectorSpecificity( const Uint32& specificity );
+	void setSelectorSpecificity( const Int64& specificity );
 
 	bool markerExists( const Uint32& marker ) const;
