+++ include/eepp/ui/css/stylesheetstyle.hpp
@@ -67,7 +67,7 @@ class EE_API StyleSheetStyle {
 
 	void setMarker( const Uint32& marker );
 
-	void setSelectorSpecificity( const Uint32& specificity );
+	void setSelectorSpecificity( const Int64& specificity );
 
 	bool updatePropertyValue( const std::string& name, const std::string& value );
